pyproject.toml
[tool.poetry]
name = "bluefn"
version = "0.1.0"
description = "toolset for bluechain"
authors = ["Samlet Wu <xiaofei.wu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fire = "^0.3.1"
typer = "^0.3.1"
mako = "^1.1.3"
py-solc-x = "^0.10.1"
web3 = "^5.12.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
bluefn = 'bluefn.__main__:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
